Convert CSV strategy editor to a tabbed screen; split the file (#505)
The strategy editor had grown to ~3000 lines in a single scrolling AlertDialog
(CreateCsvStrategyDialog). This converts it into a full navigable screen with
four tabs and breaks the file into a focused editor package.

UI:
- New Screen.CsvStrategyEditor(csvImportId, strategyId?) dispatched in
  MoneyManagerApp; CsvImportDetailScreen's "Create Strategy" and
  CsvStrategiesScreen's edit (after picking a CSV) now navigate to it instead of
  opening a dialog.
- CsvStrategyEditorScreen loads the CSV columns/rows and existing strategy from
  the repos, shows a header (back + save), a PrimaryTabRow, and the tab content:
  General, Accounts, Amount & Date, Advanced. Per-tab error markers and a
  save gate driven by validation.

Structure (new com.moneymanager.ui.screens.csvstrategy.editor package):
- CsvStrategyFormState.kt: pure extract/build round-trip + enums + helpers.
- CsvStrategyEditorState.kt: a remembered state holder mirroring all ~33 fields,
  computed validity, and toFormState(); survives tab switches.
- Tab composables (GeneralTab/AccountsTab/AmountDateTab/AdvancedTab) and the
  shared editors (ColumnDropdowns, RowConditionEditors, AccountMappingEditors,
  RuleEditors, AttributeEditors), moved out and made internal.
- CreateCsvStrategyDialog.kt deleted.

Tests: StrategyFormRoundTripTest and AttributeCandidateColumnsTest updated for the
new package; new CsvStrategyEditorStateTest asserts the holder round-trips an
extracted form state. ImportMonzoCsvE2ETest reworked to drive the tabbed screen.
